services/vault_api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pyjwt==2.8.0
cryptography==42.0.5
pydantic==2.6.1
